+++ app/(gcforms)/[locale]/(form filler)/id/[...props]/actions.ts
@@ -9,10 +9,11 @@ import { FormStatus } from "@gcforms/types";
 import { verifyHCaptchaToken } from "@lib/validation/hCaptcha";
 import { checkOne } from "@lib/cache/flags";
 import { FeatureFlags } from "@lib/cache/types";
-import { validateResponses } from "@lib/validation/validation";
 
 import { dateHasPast } from "@lib/utils";
 
+import { validateOnSubmit, validateResponses } from "@lib/validation/validation";
+import { serverTranslation } from "@root/i18n";
 import { sendNotifications } from "@lib/notifications";
 
 // Public facing functions - they can be used by anyone who finds the associated server action identifer
@@ -70,6 +71,9 @@ export async function submitForm(
       }
     }
 
+    /**
+     * This validation checks the response values against the template element types.
+     */
     const validateResponsesResult = await validateResponses(values, template);
 
     if (Object.keys(validateResponsesResult).length !== 0) {
@@ -80,6 +84,27 @@ export async function submitForm(
         )}`
       );
       // Turn this on after we've monitored the logs for a while
+      // throw new MissingFormDataError("Form data validation failed");
+    }
+
+    const { t } = await serverTranslation();
+
+    /**
+     * This validation runs the client-side validation on the server.
+     */
+    const validateOnSubmitResult = await validateOnSubmit(values, {
+      formRecord: template,
+      t: t,
+    });
+
+    if (Object.keys(validateOnSubmitResult).length !== 0) {
+      logMessage.info(
+        `[server-action][submitForm] Detected validation errors on form ${formId}. Errors: ${JSON.stringify(
+          validateOnSubmitResult
+        )}`
+      );
+      // Keeping in "passive mode" for now.
+      // Uncomment following line to throw validation error from server.
       // throw new MissingFormDataError("Form data validation failed");
     }